Move `pretty_printed_query` function to util.py
